[Lesson4] Scope
