+++ Lesson13/exercises.py
@@ -1,0 +1,26 @@
+# 1. Safe divide
+
+def safe_divide(x, y):
+    # TODO: write this function
+    return 0
+
+
+# 2. Dictionary get
+
+def dict_get(d, key):
+    # TODO: write this function
+    return None
+
+
+# 3. Strict factorial
+
+def factorial(n):
+    # TODO: raise an exception if n is negative
+    if n == 0:
+        return 1
+    return n * factorial(n - 1)
+
+
+# There's not much else here yet because we're still working on this lesson.
+# Check with a mentor if you need help understanding exceptions, or if you want
+# more examples!
