--- Lesson4/exercises.py
@@ -1,49 +1,0 @@
-# Returns the square of a number if it's odd; otherwise, returns the original number.
-def square_if_odd(num): 
-    # TODO: implement this!
-    return -1
-
-
-# Returns whether the given number is within the range of lower and upper inclusive (True or False).
-def in_range(lower, upper, num):
-    # TODO: implement this!
-    return False
-
-
-# Takes in a list of numbers and returns a list of only the even ones.
-def even_elements(nums):
-    # TODO: implement this!
-    return []
-
-
-# Takes in a string and returns the most common letter occuring in the string.
-# If there's a tie, you may return any of them.
-def most_common_letter(str):
-    # TODO: implement this!
-    return 'a'
-
-
-# Takes in a list of numbers and returns the maximum difference between any two numbers in the list.
-def max_diff(nums):
-    # TODO: implement this!
-    return -1
-
-
-# Prints an hourglass made out of '*' characters with a base the size of the number supplied * 2.
-def hourglass(size):
-    # TODO: implement this!
-    print("not implemented")
-
-
-
-
-
-
-
-
-
-# Everything below here is just to test your code. You don't need to change any
-# of this.
-
-if __name__ == '__main__':
-
