+++ Lesson5/exercises.py
@@ -1,36 +1,49 @@
-# Searches linearly through a sorted list of numbers (nums) for a number
-# (target) and returns whether or not the list contains the number.
-def linear_search(nums, target):
-    for num in nums:
-        if num == target:
-            return True
-    return False
+# Returns the square of a number if it's odd; otherwise, returns the original number.
+def square_if_odd(num): 
+    # TODO: implement this!
+    return -1
 
 
-# Uses binary search to look through a sorted list of numbers (nums) for a
-# number (target) and returns whether or not the list contains the number.
-def binary_search(nums, target):
+# Returns whether the given number is within the range of lower and upper inclusive (True or False).
+def in_range(lower, upper, num):
     # TODO: implement this!
     return False
 
 
+# Takes in a list of numbers and returns a list of only the even ones.
+def even_elements(nums):
+    # TODO: implement this!
+    return []
+
+
+# Takes in a string and returns the most common letter occuring in the string.
+# If there's a tie, you may return any of them.
+def most_common_letter(str):
+    # TODO: implement this!
+    return 'a'
+
+
+# Takes in a list of numbers and returns the maximum difference between any two numbers in the list.
+def max_diff(nums):
+    # TODO: implement this!
+    return -1
+
+
+# Prints an hourglass made out of '*' characters with a base the size of the number supplied * 2.
+def hourglass(size):
+    # TODO: implement this!
+    print("not implemented")
 
 
 
 
-# Code below here just tests if your code works properly. You don't need to
-# change any of this code.
 
 
-def test_binary_search(nums, target, desired_result):
-    if binary_search(nums, target) is desired_result:
-        print("TEST SUCCESS on: " + str(nums))
-    else:
-        print("TEST FAILED on: " + str(nums))
 
 
-test_binary_search([1, 2, 3, 6, 7, 10, 10, 10], 7, True)
-test_binary_search([1, 6, 10, 10, 10], 7, False)
-test_binary_search([], 20, False)
-test_binary_search([-80, -60, -10, 0, 12, 13], -80, True)
-test_binary_search([-80, -60, -10, 0, 12, 13], -79, False)
+
+# Everything below here is just to test your code. You don't need to change any
+# of this.
+
+if __name__ == '__main__':
+
