+++ Lesson6/maxdiff.py
