+++ Lesson9/exercises.py
@@ -1,25 +1,157 @@
-# Recursively count down from num to 0
-def count_down(num):
-    # TODO: implement this!
-    pass
+# 1. round_sum
+
+# This problem is on CodingBat; go to http://codingbat.com/prob/p179960 to work
+# on it. When you're done, come back here.
 
 
-# Recursively compute a factorial
-def factorial(num):
-	# TODO: implement this!
-	return -1
+# 2. Find last letters
+
+def find_last_letters(word):
+  result = {}
+  # TODO
+  return result
 
 
-ZOMBIE = 'z'
-CAT = 'c'
-HUMAN = 'h'
+# 3. Words by length
 
-# Flood-fill
-def flood_fill(world, x, y):
-    # Check if (x, y) is inside the world - if it's outside, do nothing.
-    if (x < 0) or (x >= len(world)):
+def words_by_length(words):
+  result = {}
+  # TODO
+  return result
+
+def word_appears(words_dict, word):
+  # TODO
+  return False
+
+def words_by_length_challenge(words):
+  result = {}
+  # TODO
+  return result
+
+
+# 4A. Splitting lists
+
+def split_list(items, count):
+  result = []
+  # TODO
+  return result
+
+def split_list_challenge(items, count):
+  result = []
+  # TODO
+  return result
+
+
+# 4B. Finding items in nested lists
+
+def which_list(lists, item):
+  # TODO
+  return -1
+
+
+# Don't forget to figure out the Big O runtime of parts A and B!
+
+
+# You don't need to change anything below here - everything below here just
+# tests whether your code works correctly. You can run this file as
+# "python exercises.py" and it will tell you if anything is wrong.
+
+
+
+
+
+
+
+
+
+
+
+
+def test_find_last_letters():
+  expected = {'c': 0, 'h': 1, 'a': 2, 'l': 4, 'e': 8, 'n': 6, 'g': 7}
+  ret = find_last_letters('challenge')
+  if ret != expected:
+    print('find_last_letters should return %r, but it returned %r' % (
+        expected, ret))
+  else:
+    print('find_last_letters is correct')
+
+def test_words_by_length():
+  animals = ['cat', 'dog', 'bird', 'sheep', 'rat', 'horse', 'human', 'lion']
+  expected = {3: ['cat', 'dog', 'rat'],
+              4: ['bird', 'lion'],
+              5: ['sheep', 'horse', 'human']}
+  ret = words_by_length(animals)
+  if ret != expected:
+    print('words_by_length(%r) should return %r, but it returned %r' % (
+        animals, expected, ret))
+    return
+  else:
+    print('words_by_length is correct')
+
+  cases = [
+    # (word, expected_result)
+    ('cat', True),
+    ('sheep', True),
+    ('capybara', False),
+    ('donut', False),  # don't be silly; this isn't an animal
+  ]
+  for word, expected_result in cases:
+    result = word_appears(expected, word)
+    if expected_result != result:
+      print('word_appears(%r) should return %r, but it returned %r' % (
+          word, expected_result, result))
+      return
+  print('word_appears is correct')
+
+  ret = words_by_length_challenge(animals)
+  if not ret:
+    return  # don't print anything if the challenge wasn't attempted
+
+  if ret != expected:
+    print('words_by_length_challenge(%r) should return %r, but it returned %r' % (
+        animals, expected, ret))
+  else:
+    print('words_by_length_challenge is correct')
+
+def test_split_list():
+  cases = [
+    # (input_list, sublist_size, expected_result)
+    ([1, 3, 5, 7], 2, [[1, 3], [5, 7]]),
+    ([1, 3, 5, 7, 9, 11, 13], 3, [[1, 3, 5], [7, 9, 11], [13]]),
+    ([1, 3, 5, 7], 10, [[1, 3, 5, 7]]),
+    ([], 3, []),
+  ]
+  for input_list, sublist_size, expected_result in cases:
+    result = split_list(input_list, sublist_size)
+    if result != expected_result:
+      print('split_list(%r, %r) should return %r, but it returned %r' % (
+          input_list, sublist_size, expected_result, result))
+      return
+  print('split_list is correct')
+
+  which_list_input = [[1, 3, 5], [5, 7, 9], [9, 11]]
+  which_list_cases = [(1, 0), (5, 0), (7, 1), (9, 1), (11, 2), (17, -1)]
+  for item, expected_result in which_list_cases:
+    result = which_list(which_list_input, item)
+    if result != expected_result:
+      print('which_list(%r, %r) should return %r, but it returned %r' % (
+          which_list_input, item, expected_result, result))
+      return
+  print('which_list is correct')
+
+  # check if the challenge was attempted
+  if split_list_challenge([1], 1):
+    for input_list, sublist_size, expected_result in cases:
+      result = split_list_challenge(input_list, sublist_size)
+      if result != expected_result:
+        print('split_list_challenge(%r, %r) should return %r, but it returned %r' % (
+            input_list, sublist_size, expected_result, result))
         return
-    if (y < 0) or (y >= len(world[x])):
-        return
+    print('split_list_challenge is correct')
 
-    # TODO: implement the rest of this!
+
+if __name__ == '__main__':
+  test_find_last_letters()
+  test_words_by_length()
+  test_split_list()
