+++ README.md
@@ -9,15 +9,16 @@ Welcome to StreetCode's Next Level Programming class! In this class we will go f
 1. [Some Housekeeping and Getting Started with Python](Lesson1)
 2. [Getting Familiar with Python](Lesson2)
 3. [Lists and Dictionaries](Lesson3)
-4. [Let's get some practice](Lesson4)
-5. [Searching and an Introduction to Algorithmic Analysis](Lesson5)
-6. [More Practice](Lesson6)
-7. [An Introduction to Sorting](Lesson7)
-8. [Data Structures and Algorithms Practice](Lesson8)
-9. [Introduction to Recursion](Lesson9)
-10. [Recursion Practice](Lesson10)
-11. [Object-Oriented Programming (Classes)](Lesson11)
-12. [Exceptions and Advanced Control Flow](Lesson12)
+4. [Scope](Lesson4)
+5. [Let's get some practice](Lesson5)
+6. [Searching and an Introduction to Algorithmic Analysis](Lesson6)
+7. [More Practice](Lesson7)
+8. [An Introduction to Sorting](Lesson8)
+9. [Data Structures and Algorithms Practice](Lesson9)
+10. [Introduction to Recursion](Lesson10)
+11. [Recursion Practice](Lesson11)
+12. [Object-Oriented Programming (Classes)](Lesson12)
+13. [Exceptions and Advanced Control Flow](Lesson13)
 
 ## Prerequisites for taking this class
